requirements/base.txt
google-generativeai>=0.7.0
websockets>=12.0
pyyaml>=6.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
jsonschema>=4.23.0
numpy>=2.1.0
